Mazi pataisymai

+++ src/direct_method/DirectMethodCalculator.java
@@ -7,12 +7,12 @@ import structures.Vector;
 public class DirectMethodCalculator {
 
     // Paduodamas tiesinis abeceles kodas - vektoriu masyvas, grazinamas svoriu skirstinys
-    public int[] calculateDistribution(Vector[] code) throws Exception {
+    public long[] calculateDistribution(Vector[] code) throws Exception {
         if (code.length == 0) {
             throw new Exception("Kodas tuscias");
         }
         int k = code[0].getSize();
-        int[] distribution = new int[k + 1];
+        long[] distribution = new long[k + 1];
         for (Vector word : code) {
             int weight = calculateWeight(word);
             distribution[weight]++;
